Removed worker/celery layer and refactored into separate BigQC app. Removed associated tests, dependencies, ci/cd, xstream configuration. This package now comprises just the web layer of the whole stack. Using coltonbh/bigqc-worker:latest for all local development and CI/CD testing.
