+++ .circleci/config.yml
@@ -16,10 +16,6 @@ parameters:
     type: string
     default: terachem-cloud
 
-  tcc_worker_docker_repo:
-    type: string
-    default: terachem-cloud-worker
-
   server_stacks_directory:
     type: string
     default: stacks
@@ -54,10 +50,10 @@ jobs:
       - image: circleci/python:3.9
       - image: rabbitmq:3.8-alpine
       - image: redis:6-alpine
-      - image: mtzgroup/terachem-cloud-worker:testing
+      - image: coltonbh/bigqc-worker
         environment:
           - C_FORCE_ROOT=true # to run celery as root with pickle serializer
-        command: celery -A terachem_cloud.workers.tasks worker -Q celery,private_queue --without-heartbeat --without-mingle --without-gossip --loglevel=INFO
+        command: celery -A bigqc.tasks worker -Q celery,private_queue --without-heartbeat --without-mingle --without-gossip --loglevel=INFO
     steps:
       - checkout
       - restore_cache:
@@ -211,7 +207,6 @@ jobs:
       - run:
           name: Push docker-compose files to server
           command: |
-            scp docker/docker-compose.base.yaml "root@${MTZLAB_SWARM_HOST}:${SERVER_FILEPATH}/docker-compose.base.yaml"
             scp ${DOCKER_COMPOSE_FILENAME} "root@${MTZLAB_SWARM_HOST}:${SERVER_FILEPATH}/${DOCKER_COMPOSE_FILENAME}"
       - run:
           name: Login server to Martinez Group Docker Repo
@@ -220,54 +215,7 @@ jobs:
             ssh root@${MTZLAB_SWARM_HOST} ESCAPED_DOCKERHUB_PASSWORD=${ESCAPED_DOCKERHUB_PASSWORD} DOCKERHUB_USERNAME=${DOCKERHUB_USERNAME} "echo ${ESCAPED_DOCKERHUB_PASSWORD} | docker login -u ${DOCKERHUB_USERNAME} --password-stdin" #  pragma: allowlist secret
       - run:
           name: Deploy docker stack on server
-          command: ssh root@${MTZLAB_SWARM_HOST} "docker stack deploy --prune --with-registry-auth -c ${SERVER_FILEPATH}/docker-compose.base.yaml -c ${SERVER_FILEPATH}/${DOCKER_COMPOSE_FILENAME} tcc"
-
-  deploy_workers:
-    parameters:
-      stage:
-        description: dev or prod
-        type: string
-    docker:
-      - image: cimg/base:2020.01
-        environment:
-          DOCKER_COMPOSE_FILENAME: docker-compose.xstream.<< parameters.stage >>.yaml
-          SERVER_FILEPATH: << pipeline.parameters.server_stacks_directory >>/<< parameters.stage >>/tcc
-          XSTREAM_LOGIN_HOST: xstream-login.stanford.edu
-    steps:
-      - checkout
-      - add_ssh_keys:
-          fingerprints:
-            - 94:89:ed:6f:77:06:0e:a1:66:73:b6:36:2c:42:9c:7d
-      - run:
-          name: Install kinit
-          command: sudo apt-get update && sudo apt-get install -y krb5-user
-      - run:
-          name: Configure kinit for Stanford
-          command: sudo curl -o /etc/krb5.conf https://web.stanford.edu/dept/its/support/kerberos/dist/krb5.conf
-      - full_image_name:
-          docker_repo: << pipeline.parameters.tcc_worker_docker_repo >>
-      - run:
-          name: Fill docker-compose template with context
-          command: |
-            eval "echo \"$(cat $(pwd)/docker/${DOCKER_COMPOSE_FILENAME})\"" > ${DOCKER_COMPOSE_FILENAME}
-      - run:
-          name: Login to XStream with kinit
-          command: |
-            ssh-keyscan -H ${XSTREAM_LOGIN_HOST} >> ~/.ssh/known_hosts
-            echo ${STANFORD_PW} | kinit ${STANFORD_UN}@stanford.edu
-      - run:
-          name: Make stack directory on server
-          command: |
-            ssh -o ProxyCommand="ssh -o GSSAPIAuthentication=yes -o GSSAPIDelegateCredentials=yes ${STANFORD_UN}@xstream-login.stanford.edu nc xs7-0001 22" -o StrictHostKeyChecking=no ${STANFORD_UN}@xs7-0001 "mkdir -p ${SERVER_FILEPATH}"
-      - run:
-          name: Push docker-compose files to server
-          command: |
-            scp -o ProxyCommand="ssh -o GSSAPIAuthentication=yes -o GSSAPIDelegateCredentials=yes ${STANFORD_UN}@xstream-login.stanford.edu nc xs7-0001 22" -o StrictHostKeyChecking=no ${DOCKER_COMPOSE_FILENAME} ${STANFORD_UN}@xs7-0001:${SERVER_FILEPATH}/${DOCKER_COMPOSE_FILENAME}
-      - run:
-          name: Deploy docker stack on server
-          command: |
-            ESCAPED_DOCKERHUB_PASSWORD=`printf "%q" ${DOCKERHUB_PASSWORD}` #  pragma: allowlist secret
-            ssh -o ProxyCommand="ssh -o GSSAPIAuthentication=yes -o GSSAPIDelegateCredentials=yes ${STANFORD_UN}@xstream-login.stanford.edu nc xs7-0001 22" -o StrictHostKeyChecking=no ${STANFORD_UN}@xs7-0001 ESCAPED_DOCKERHUB_PASSWORD=${ESCAPED_DOCKERHUB_PASSWORD} DOCKERHUB_USERNAME=${DOCKERHUB_USERNAME} "echo ${ESCAPED_DOCKERHUB_PASSWORD} | docker login -u ${DOCKERHUB_USERNAME} --password-stdin && docker stack deploy --prune --with-registry-auth -c ${SERVER_FILEPATH}/${DOCKER_COMPOSE_FILENAME} << parameters.stage >>-tcc-worker" #  pragma: allowlist secret
+          command: ssh root@${MTZLAB_SWARM_HOST} "docker stack deploy --prune --with-registry-auth -c ${SERVER_FILEPATH}/${DOCKER_COMPOSE_FILENAME} tcc"
 
 workflows:
   main:
@@ -310,23 +258,6 @@ workflows:
             tags:
               only: /.*/ # required since prod_web_deploy requires this step and has tag filters
 
-      # Build Workers
-      - build_workers_approval:
-          type: approval
-          filters:
-            tags:
-              only: /.*/ # required since prod_workers_deploy requires this step and has tag filters
-      - build:
-          name: worker_build
-          requires:
-            - test_code
-            - build_workers_approval
-          docker_repo: terachem-cloud-worker
-          dockerfile_name: celeryworker.dockerfile
-          filters:
-            tags:
-              only: /.*/ # required since prod_workers_deploy requires this step and has tag filters
-
       # Deploy dev
       - dev_deploy_web_approval:
           type: approval
@@ -343,21 +274,6 @@ workflows:
           requires:
             - dev_deploy_web_approval
             - web_build
-          stage: dev
-
-      - dev_deploy_workers_approval:
-          type: approval
-          filters:
-            tags:
-              only: /.*/
-      - deploy_workers:
-          name: dev_workers_deploy
-          filters:
-            tags:
-              only: /.*/
-          requires:
-            - dev_deploy_workers_approval
-            - worker_build
           stage: dev
 
       # Prod deploy
@@ -380,23 +296,4 @@ workflows:
           requires:
             - prod_deploy_web_approval
             - web_build
-          stage: prod
-
-      - prod_deploy_worker_approval:
-          type: approval
-          filters: # Only run on tagged branches
-            tags:
-              only: /\d*\.\d*\.\d*/
-            branches:
-              ignore: /.*/
-      - deploy_workers:
-          name: prod_workers_deploy
-          filters:
-            tags:
-              only: /\d*\.\d*\.\d*/
-            branches:
-              ignore: /.*/
-          requires:
-            - prod_deploy_worker_approval
-            - worker_build
           stage: prod
