+++ Pipfile
@@ -12,7 +12,6 @@ pytest = "*"
 pre-commit = "*"
 pytest-cov = "*"
 requests = "*"
-pytest-celery = "*"
 pytest-timeout = "*"
 autoflake = "*"
 types-requests = "*"
@@ -28,14 +27,11 @@ uvloop = "*"
 httptools = "*"
 typing-extensions = "*"
 aiofiles = "*"
-pyberny = "*"
-qcengine = {ref = "feature-terachem-frontend-harness", git = "https://github.com/coltonbh/QCEngine.git"}
-geometric = {ref = "bugfix-no-dirname-for-run-json", git = "https://github.com/coltonbh/geomeTRIC.git"}
-tcpb = ">=0.11.0"
-celery = {extras = ["redis"], version = "*"}
+tcpb = ">=0.12.0"
 cached-property = "*"
 qcelemental = "==0.24.0"
 fastapi = "==0.65.1"
+bigqc = {ref = "0.1.1", git = "https://github.com/coltonbh/bigqc.git"}
 
 [requires]
 python_version = "3.9"
@@ -45,5 +41,5 @@ allow_prereleases = true
 
 [scripts]
 tests = "bash ./scripts/test.sh"
-start = "docker-compose --env-file .env -f docker/docker-compose.base.yaml -f docker/docker-compose.local.yaml up -d --build"
-stop = "docker-compose -f docker/docker-compose.base.yaml -f docker/docker-compose.local.yaml down"
+start = "docker-compose --env-file .env -f docker/docker-compose.local.yaml up -d --build web-server worker"
+stop = "docker-compose -f docker/docker-compose.local.yaml down"
