+++ README.md
@@ -36,10 +36,10 @@ pipenv run stop
 For more granularity you can use docker to run various components of the service. If you prefer to run a non-dockerized version of the web app and/or the celery worker those commands can be found below. Useful commands to get started running sub-components of the application:
 
 ```sh
-docker-compose -f docker/docker-compose.base.yaml -f docker/docker-compose.local.yaml up -d --build [service_of_interest]
+docker-compose -f docker/docker-compose.local.yaml up -d --build [services_of_interest]
 ```
 
-Run non-dockerized local celery worker
+Run non-dockerized local bigqc worker
 
 ```sh
 pipenv run celery -A terachem_cloud.workers.tasks worker --loglevel=INFO
@@ -52,12 +52,13 @@ pipenv run uvicorn terachem_cloud.main:app --reload
 ```
 
 ### Development on Fire (or any machine with GPUs)
-Developing on a machine with GPUs means you can run `TeraChem` in server mode inside the docker worker. For development purposes, it's easiest to startup the `docker-compose` stack sans worker and then run the worker with `docker run` directly since `docker-compose` lacks good GPU support.
 
-The additional `docker-compose.fire.yaml` file places all services on an external `tcc` network. Then when starting terachem via the docker command below it is also added to the network. The .env variable `TERACHEM_PBS_HOST` must be set to the name of the terachem container (named `terachem` below)
+Developing on a machine with GPUs means you can run `TeraChem` in server mode and the `BigQC` worker can send work to it. Simply include `terachem` and `file-server` in the list of `services_of_interest` in the `docker-compose` commands noted above.
+
+If you can't run a modern version of `docker-compose` that has good GPU support, you can run the `TeraChem` worker as a separation container and network it to the `TeraChem Cloud` stack. The additional `docker-compose.fire.yaml` file places all services on an external `tcc` network. Then when starting terachem via the docker command below it is also added to the network. The .env variable `TERACHEM_PBS_HOST` must be set to the name of the terachem container (named `terachem` below)
 
 ```sh
-docker-compose -f docker/docker-compose.base.yaml -f docker/docker-compose.local.yaml -f docker/docker-compose.fire.yaml up -d --build web-server mq redis worker
+docker-compose -f docker/docker-compose.local.yaml -f docker/docker-compose.fire.yaml up -d --build web-server mq redis worker
 
 docker run -d --rm -v terachem-scratch:/scratch -v /home/coltonbh/license.key:/terachem/license.key -p 11111:11111 --gpus '"device=0,1"' --network="tcc" --name terachem mtzgroup/terachem:1.9-2021.12-dev-arch-sm_52-sm_80 && docker logs terachem -f
 
@@ -96,9 +97,9 @@ A test summary will be output to `/htmlcov`. Open `/htmlcov/index.html` to get a
 - In the directory on the server from which the `docker-compose.web.yaml` file will deploy, create the following files and populate with their correct secrets:
   - server.env
     - `BASE_URL=https://yourdomain.com`
-    - `CELERY_BROKER_CONNECTION_STRING=amqp://${USERNAME}:${PASSWORD}@mq:5672` # pragma: allowlist secret
+    - `BIGQC_BROKER_URL=amqp://${USERNAME}:${PASSWORD}@mq:5672` # pragma: allowlist secret
       - Note that the host should correspond to the service name for the `rabbitmq` instance found in the `docker-compose.web.yaml` file since we want the web server to connect locally rather than over the open internet. Also, note we are connecting to the port on which `rabbitmq` is running _insecurely_ on the stack. Traefik is providing TLS termination for external services that must connect _securely_ to `rabbit`.
-    - `CELERY_BACKEND_CONNECTION_STRING=redis://:${PASSWORD}@redis:6379/0`
+    - `BIGQC_BACKEND_URL=redis://:${PASSWORD}@redis:6379/0`
       - See note above for `rabbitmq`. Same logic applies.
     - `AUTH0_DOMAIN=xxx`
     - `AUTH0_CLIENT_ID=xxx`
@@ -113,7 +114,7 @@ A test summary will be output to `/htmlcov`. Open `/htmlcov/index.html` to get a
 ### Workers
 
 - In the directory on the server from which the `docker-compose.worker.yaml` file will deploy, create the following files and populate with their correct secrets:
-  - `CELERY_BROKER_CONNECTION_STRING=amqps://${USERNAME}:${PASSWORD}@rmq.dev.mtzlab.com:5671`
-    - Note the amqpS protocol. Since we are connecting over the open internet we require TLS. Also note the use of the `5671` *secure* port for `amqps` connections. This is the port on which `traefik` is listening for `amqps` connections.
-  - `CELERY_BACKEND_CONNECTION_STRING=rediss://:${PASSWORD}@redis.dev.mtzlab.com:6379/0?ssl_cert_reqs=CERT_NONE`
+  - `BIGQC_BROKER_URL=amqps://${USERNAME}:${PASSWORD}@rmq.dev.mtzlab.com:5671`
+    - Note the amqpS protocol. Since we are connecting over the open internet we require TLS. Also note the use of the `5671` _secure_ port for `amqps` connections. This is the port on which `traefik` is listening for `amqps` connections.
+  - `BIGQC_BACKEND_URL=rediss://:${PASSWORD}@redis.dev.mtzlab.com:6379/0?ssl_cert_reqs=CERT_NONE`
     - Same logic for the `rabbitmq` connection string applies here. Note that we do not verify the SSL certificate. This is because `traefik` is dynamically generating and renewing SSL certificates so we do not have a "permanent" certificate that we can place on the client and use for verification. Very low risk of man-in-the-middle attacks here.
