--- docker/docker-compose.base.yaml
@@ -1,12 +1,0 @@
-version: "3.8"
-
-services:
-  mq:
-    image: rabbitmq:3.8-management-alpine
-    # Must specify hostname since data is stored based on hostname and we don't want a random, docker-assigned
-    # hostname with each new restart of the service. This would result in any past messages stored by the
-    # former container in the volume to be unseen by the new container.
-    # https://hub.docker.com/_/rabbitmq/ under "How to use this image"
-    hostname: rmq-host1
-  redis:
-    image: redis:6-alpine
