+++ docker/docker-compose.web.yaml
@@ -22,59 +22,6 @@ services:
         - traefik.http.routers.tcc-web-server.tls=true
         - traefik.http.routers.tcc-web-server.tls.certresolver=le
         - traefik.http.services.tcc-web-server-service.loadbalancer.server.port=8000
-
-  mq:
-    networks:
-      - traefik-public
-    env_file: rabbit.env
-    volumes:
-      - rabbitmq:/var/lib/rabbitmq
-      - ./rabbitmq.conf:/etc/rabbitmq/rabbitmq.conf
-    deploy:
-      labels:
-        - traefik.enable=true
-        - traefik.docker.network=traefik-public
-        - traefik.constraint-label=traefik-public
-
-        # RabbitMQ
-        - traefik.tcp.routers.tcc-rabbitmq.rule=HostSNI(\`*\`)
-        - traefik.tcp.routers.tcc-rabbitmq.entrypoints=rabbitmq-secure
-        - traefik.tcp.routers.tcc-rabbitmq.service=tcc-rabbitmq-service
-        - traefik.tcp.routers.tcc-rabbitmq.tls=true
-        - traefik.tcp.routers.tcc-rabbitmq.tls.certresolver=le
-        - traefik.tcp.services.tcc-rabbitmq-service.loadbalancer.server.port=5672
-
-        # MGMT Console
-        - traefik.http.routers.tcc-rabbitmq-console.rule=Host(${MQ_CONSOLE_TRAEFIK_HOST})
-        - traefik.http.routers.tcc-rabbitmq-console.entrypoints=https
-        - traefik.http.routers.tcc-rabbitmq-console.tls=true
-        - traefik.http.routers.tcc-rabbitmq-console.tls.certresolver=le
-        - traefik.http.routers.tcc-rabbitmq-console.service=tcc-rabbitmq-console-service
-        - traefik.http.services.tcc-rabbitmq-console-service.loadbalancer.server.port=15672
-  redis:
-    networks:
-      - traefik-public
-    volumes:
-      - redis:/data
-      - ./redis.conf:/etc/redis/redis.conf
-    deploy:
-      labels:
-        - traefik.enable=true
-        - traefik.docker.network=traefik-public
-        - traefik.constraint-label=traefik-public
-
-        - traefik.tcp.routers.tcc-redis.rule=HostSNI(\`*\`)
-        - traefik.tcp.routers.tcc-redis.entrypoints=redis-secure
-        - traefik.tcp.routers.tcc-redis.tls=true
-        - traefik.tcp.routers.tcc-redis.tls.certresolver=le
-        - traefik.tcp.routers.tcc-redis.service=tcc-redis-service
-        - traefik.tcp.services.tcc-redis-service.loadbalancer.server.port=6379
-    command: redis-server /etc/redis/redis.conf
-
-volumes:
-  rabbitmq:
-  redis:
-
 networks:
   traefik-public:
     external: true
