--- docker/docker-compose.xstream.dev.yaml
@@ -1,160 +1,0 @@
-# A bash-able template that will fill with variables if correct environment variables are loaded
-# source context.{dev | prod}
-# eval echo "" > output.yaml
-
-# XStream dev worker configuration
-# For notes on setup see docs/development/architecture.md#XStream-Architecture-Setup
-
-version: \"3.7\" #XStream can only support up to 3.7. Using docker 18.06.1-ce
-
-#######################################
-##### DEFAULT IMAGE CONFIGURATION #####
-#######################################
-
-# Default images and setup for worker containers
-x-default-worker: &default-worker
-  image: ${FULL_IMAGE_NAME}
-  env_file: worker.env
-  environment:
-    - TERACHEM_PBS_PORT=11111
-    - C_FORCE_ROOT=true # to run celery as root with pickle serializer
-
-# Default images and setup for terachem containers
-x-default-terachem: &default-terachem
-  image: mtzgroup/terachem:1.9-2022.03-dev-cuda10.2-sm_37-sm_52-sm_60
-  secrets:
-    - source: terachem_license
-      target: /terachem/license.key
-  volumes:
-    - scratch-vol:/scratch
-
-# Default images and setup for TeraChem Frontend container
-x-default-terachem-frontend: &default-terachem-frontend
-  image: mtzgroup/terachem-frontend:0.2.0-noproxy
-  environment:
-    - AUTOINDEX_FORMAT=json
-    - WORKER_PROCESSES=1
-  volumes:
-    - scratch-vol:/usr/share/nginx/html
-
-#######################################
-#### PER NODE DEPLOY CONFIGURATION ####
-#######################################
-
-# NOTE: No xs7-0001 workers accoriding to Stephane to maintain cluster stability.
-# Does this really matter? Not clear.
-
-# Deploy to xs7-0002 node
-x-xs7-0002-deploy: &xs7-0002
-  placement:
-    constraints:
-      - "node.hostname==xs7-0002"
-
-# Deploy to xs7-0003 node
-x-xs7-0003-deploy: &xs7-0003
-  placement:
-    constraints:
-      - "node.hostname==xs7-0003"
-
-# Deploy to xs7-0100 node
-x-xs7-0100-deploy: &xs7-0100
-  placement:
-    constraints:
-      - "node.hostname==xs7-0100"
-
-#######################################
-#### SERVICE-SPECIFIC CONFIGURATION ###
-#######################################
-
-# Celery workers on xs7-0002
-x-xs7-0002-worker: &xs7-0002-worker
-  <<: *default-worker
-  deploy:
-    <<: *xs7-0002
-
-# TeraChem servers on xs7-0002
-x-xs7-0002-terachem: &xs7-0002-terachem
-  <<: *default-terachem
-  deploy:
-    <<: *xs7-0002
-
-# Celery workers on xs7-0003
-x-xs7-0003-worker: &xs7-0003-worker
-  <<: *default-worker
-  deploy:
-    <<: *xs7-0003
-
-# TeraChem servers on xs7-0003
-x-xs7-0003-terachem: &xs7-0003-terachem
-  <<: *default-terachem
-  deploy:
-    <<: *xs7-0003
-
-# Celery workers on xs7-0100
-x-xs7-0100-worker: &xs7-0100-worker
-  <<: *default-worker
-  deploy:
-    <<: *xs7-0100
-
-# TeraChem servers on xs7-0100
-x-xs7-0100-terachem: &xs7-0100-terachem
-  <<: *default-terachem
-  deploy:
-    <<: *xs7-0100
-
-#######################################
-######## SERVICES SPECIFICATION #######
-#######################################
-services:
-  # Pod 2: xs7-0002, GPU 0
-  xs7-0002-worker-0:
-    <<: *xs7-0002-worker
-    environment:
-      - TERACHEM_PBS_HOST=xs7-0002-terachem-0
-      - TERACHEM_FE_HOST=xs7-0002-terachem-frontend
-  xs7-0002-terachem-0:
-    <<: *xs7-0002-terachem
-    environment:
-      - NVIDIA_VISIBLE_DEVICES=0
-  xs7-0002-terachem-frontend:
-    <<: *default-terachem-frontend
-    deploy:
-      <<: *xs7-0002
-
-  # Pod 3: xs7-0003, GPU 0
-  xs7-0003-worker-0:
-    <<: *xs7-0003-worker
-    environment:
-      - TERACHEM_PBS_HOST=xs7-0003-terachem-0
-      - TERACHEM_FE_HOST=xs7-0003-terachem-frontend
-  xs7-0003-terachem-0:
-    <<: *xs7-0003-terachem
-    environment:
-      - NVIDIA_VISIBLE_DEVICES=0
-  xs7-0003-terachem-frontend:
-    <<: *default-terachem-frontend
-    deploy:
-      <<: *xs7-0003
-
-  # Pod 3: xs7-0100, GPU 0,1
-  xs7-0100-worker-0:
-    <<: *xs7-0100-worker
-    environment:
-      - TERACHEM_PBS_HOST=xs7-0100-terachem-0
-      - TERACHEM_FE_HOST=xs7-0100-terachem-frontend
-    command: celery -A terachem_cloud.workers.tasks worker -Q celery,p100 --without-heartbeat --without-mingle --without-gossip --loglevel=INFO
-  xs7-0100-terachem-0:
-    <<: *xs7-0100-terachem
-    environment:
-      - NVIDIA_VISIBLE_DEVICES=0
-  xs7-0100-terachem-frontend:
-    <<: *default-terachem-frontend
-    deploy:
-      <<: *xs7-0100
-
-secrets:
-  terachem_license:
-    external: true
-
-volumes:
-  scratch-vol:
