+++ scripts/test.sh
@@ -1,10 +1,10 @@
-set -e
+set -x
 
 # Start docker redis, rabbitmq, and psi4 celery worker containers
-docker-compose --env-file .env -f docker/docker-compose.base.yaml -f docker/docker-compose.local.yaml up -d --build mq redis worker
+docker-compose --env-file .env -f docker/docker-compose.local.yaml up -d --build mq redis worker
 
 # Run tests
 pytest --cov-report html:htmlcov --cov
 
 # Stop docker containers
-docker-compose -f docker/docker-compose.base.yaml -f docker/docker-compose.local.yaml down --remove-orphans
+docker-compose -f docker/docker-compose.local.yaml down --remove-orphans
